# techforce-pro/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: FieldCommand — Premium fire protection field service management platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /employees/{id}/schedule:
    get:
      operationId: getEmployeeSchedule
      tags: [employees]
      summary: Get employee jobs for a date range
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"

  # ── CUSTOMERS ────────────────────────────────────────────────────────────────
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List all appointments
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    put:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete an appointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── OPENAI ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, name, role, certType, salary, billableRate, shopDaysAllowed, shopDaysUsedYtd, utilizationPct, status]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        certType:
          type: string
          enum: [suppression, sprinkler, extinguisher, helper, admin]
        salary:
          type: number
        billableRate:
          type: number
        shopDaysAllowed:
          type: integer
        shopDaysUsedYtd:
          type: integer
        allowedTrainingDays:
          type: integer
        trainingDaysUsedYtd:
          type: integer
        utilizationPct:
          type: number
        status:
          type: string
          enum: [active, inactive]
        hireDate:
          type: string
        notes:
          type: string
    EmployeeInput:
      type: object
      required: [name, role, certType, salary, billableRate]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        certType:
          type: string
          enum: [suppression, sprinkler, extinguisher, helper, admin]
        salary:
          type: number
        billableRate:
          type: number
        shopDaysAllowed:
          type: integer
        hireDate:
          type: string
        notes:
          type: string
    EmployeeUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        certType:
          type: string
        salary:
          type: number
        billableRate:
          type: number
        shopDaysAllowed:
          type: integer
        shopDaysUsedYtd:
          type: integer
        allowedTrainingDays:
          type: integer
        trainingDaysUsedYtd:
          type: integer
        utilizationPct:
          type: number
        status:
          type: string
        notes:
          type: string
...
    Job:
      type: object
      required: [id, customerId, customerName, employeeId, techName, serviceType, scheduledDate, scheduledTime, status, priority, revenue, certRequired]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        employeeId:
          type: ["integer", "null"]
        techName:
          type: string
        serviceType:
          type: string
        description:
          type: string
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, return, reschedule, emergency, cancelled, non_compliant, will_return]
        priority:
          type: string
          enum: [low, medium, high, critical]
        revenue:
          type: number
        certRequired:
          type: string
        recurring:
          type: boolean
        recurringInterval:
          type: ["string", "null"]
        notes:
          type: string
        completedAt:
          type: ["string", "null"]
        invoiceId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        nonComplianceReason:
          type: ["string", "null"]
        nonComplianceNotifiedAt:
          type: ["string", "null"]
...
    Appointment:
      type: object
      required: [id, title, date, startTime, type, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        type:
          type: string
          enum: [meeting, site-visit, call, internal, other]
        participants:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdBy:
          type: ["string", "null"]
        createdAt:
          type: string
    CreateAppointmentBody:
      type: object
      required: [title, date, startTime, type]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        type:
          type: string
          enum: [meeting, site-visit, call, internal, other]
        participants:
          type: string
        location:
          type: string
        notes:
          type: string
        createdBy:
          type: string
    UpdateAppointmentBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        type:
          type: string
        participants:
          type: string
        location:
          type: string
        notes:
          type: string
...